گزارش تراز کالا فروش

+++ ApiKarbord/Models/Public/Web_CGru.cs
@@ -17,15 +17,15 @@
 
         public int Level { get; set; }
 
-        public long? Code1 { get; set; }
+        public int? Code1 { get; set; }
 
-        public long? Code2 { get; set; }
+        public int? Code2 { get; set; }
 
-        public long? Code3 { get; set; }
+        public int? Code3 { get; set; }
 
-        public long? Code4 { get; set; }
+        public int? Code4 { get; set; }
 
-        public long? Code5 { get; set; }
+        public int? Code5 { get; set; }
 
         public short? Mode { get; set; }
     }
